move JourneyCard back into JourneyList
