+++ apps/journeys-admin/src/components/JourneyList/JourneyCard/JourneyCard.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react'
 import JourneyCard from './JourneyCard'
-import { defaultJourney } from '../JourneyList/journeyListData'
+import { defaultJourney } from '../journeyListData'
 
 describe('JourneyCard', () => {
   it('should render Journey Card', () => {
